+++ devel/0140.md
@@ -64,6 +64,39 @@ object->port 区块本身是内部打印引擎（77 个 `*_to_port` 函数，无
 - pos_int_to_str 系列等小工具可随迁
 - eq/equal 区块（33975）仅用 clear_shared_info，导出声明即可
 
+## 2026-08-26 第三步：迁移 object->port 主体（第二步提交）
+
+### What
+1. 将 s7.c object->port 区块（2948 行：display_functions 分发表、约 50 个 *_to_port
+   类型打印函数、object_out/cyclic_out 主循环、closure 可读打印等 77 个函数）
+   迁移到 s7_scheme_write.c（write.c 现 3738 行，s7.c 从 85940 降至 82545 行）
+2. s7.c 导出 20 个原 static 函数供 write.c 使用（声明在 s7_internal.h）：
+   liberate（改 extern inline，与 mallocate_block 同模式）、pos_int_to_str 系列、
+   wrap_real、set_dlist_1、type_name、tree_is_cyclic、integer_to_string 系列、
+   floatify、insert_spaces、number_to_string_base_10、resize_port_data、
+   string_write_string_resized、port_write_unicode_char、hash_table_iterate、
+   port_write_vector_typer、hash_table_checker_name、print_debugging_state
+3. s7.c 导出 dignum / chars / t_number_p 全局表（extern 声明在 s7_internal.h）
+4. write.c 导出 6 个入口（声明在 s7_scheme_write.h）：circular_list_entries、
+   closure_name、find_closure、find_typer、hash_table_typer_name、init_display_functions
+5. s7i_object_out 桥接函数一并迁入 write.c；s7.c 中 3 处直接调用 object_out 的
+   （s7_object_to_c_string / s7_object_to_string / g_object_to_string）改走桥接
+6. write.c 新增 include s7_scheme_let.h（let_ref_p_pp）和 s7_continuation.h
+   （call_exit_name / continuation_to_port）
+7. s7_internal.h 补充镜像宏与条件块：resize_stack / T_Itr_Pos / T_Itr_Let 带
+   S7_DEBUGGING 条件分支，c_object_print 带 !DISABLE_DEPRECATED 守卫
+
+### How
+- 与上一步相同的编译错误驱动固定点迭代收集宏；非宏符号（函数、全局表）逐个
+  去 static 并加声明
+- 注意点：s7.c 内若有 static 前向声明（如 type_name 4398 行），只去掉定义处的
+  static 会导致前向声明"static 覆盖"而链接失败，前向声明也要一并处理
+
+### 验证
+- xmake b goldfish 编译链接通过
+- bin/gf test --all：1515 个测试全部通过（timeit 偶发失败为计时敏感，复跑通过）
+- write/display/write-shared/object->string/format 等 12 个打印相关测试单独复跑全部通过
+
 ## 2026-08-26 第二步：迁移 cycles/shared-info 机制（第一步提交）
 
 ### What
